requirements.txt
kafka-python==2.0.2
requests==2.26.0
redis==3.5.3
python-dotenv==0.19.0